docker/exos-correction/exo5/correction/requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.26